Fix website loading errors
- Install missing Next.js dependencies
- Fix import path for Layout component in MainLayout.tsx
- Fix duplicate Calendar name conflict in calendar.tsx
- Move components from pages/components to proper components directory
- Simplify Next.js configuration to resolve development server issues
- Downgrade Next.js to stable version 14.2.5
- Site now loads successfully in production mode

Resolves console errors and ensures proper site functionality.
