+++ package.json
@@ -228,11 +228,8 @@
     "optimize:performance": "node scripts/performance-optimizer.cjs",
     "audit:security": "node scripts/security-audit-enhanced.cjs",
     "optimize:ux": "node scripts/ux-optimizer.cjs",
-    "improve:all": "node automation/comprehensive-app-improvement.cjs",
-    "improve:all": "node scripts/quick-app-improvements.cjs run",
     "improve:all": "bash scripts/comprehensive-app-improvements.sh",
     "sanitize:data": "node tools/sanitize-data-files.cjs",
-    "automation:safe": "node tools/run-safe-automations.cjs",
     "automation:safe": "bash scripts/run-safe-automations.sh",
     "ci:smoke": "npm run build && npm run test:smoke",
     "ci:consolidated": "bash scripts/ci-consolidated.sh",
@@ -266,13 +263,7 @@
   "dependencies": {
     "@google-analytics/data": "^5.2.0",
     "@google/generative-ai": "^0.24.1",
-    "@supabase/ssr": "^0.6.1",
-    "@supabase/supabase-js": "^2.39.0",
     "@hookform/resolvers": "^5.2.1",
-    "@modelcontextprotocol/sdk": "^1.17.1",
-    "@modelcontextprotocol/server-filesystem": "^2025.7.29",
-    "@modelcontextprotocol/server-memory": "^2025.8.4",
-    "@modelcontextprotocol/server-sequential-thinking": "^2025.7.1",
     "@radix-ui/react-accordion": "^1.1.2",
     "@radix-ui/react-alert-dialog": "^1.0.5",
     "@radix-ui/react-avatar": "^1.0.4",
@@ -300,6 +291,7 @@
     "@radix-ui/react-tooltip": "^1.0.7",
     "@reduxjs/toolkit": "^2.9.0",
     "@stripe/stripe-js": "^7.9.0",
+    "@supabase/ssr": "^0.6.1",
     "@supabase/supabase-js": "^2.57.0",
     "@tailwindcss/postcss": "^4.1.12",
     "@tanstack/react-query": "^5.85.5",
@@ -312,19 +304,6 @@
     "clsx": "^2.1.1",
     "critters": "^0.0.23",
     "cron-parser": "^5.3.0",
-    "framer-motion": "^12.23.12",
-    "fs-extra": "^11.3.1",
-    "glob": "^11.0.3",
-    "import-fresh": "^3.3.1",
-    "lucide-react": "^0.536.0",
-    "next": "^14.0.0",
-    "node-cron": "^4.2.1",
-    "openai": "^5.11.0",
-    "postcss": "^8.4.32",
-    "react": "^18.0.0",
-    "react-datepicker": "^8.4.0",
-    "react-dom": "^18.0.0",
-    "uuid": "^11.1.0",
     "d3-color": "^3.1.0",
     "d3-format": "^3.1.0",
     "d3-path": "^3.1.0",
@@ -340,10 +319,11 @@
     "import-fresh": "^3.3.1",
     "lucide-react": "^0.536.0",
     "mongodb": "^6.19.0",
-    "next": "^15.5.2",
+    "next": "^14.2.5",
     "node-cron": "^4.2.1",
     "openai": "^5.11.0",
     "path": "^0.12.7",
+    "postcss": "^8.4.32",
     "postcss-preset-env": "^10.3.1",
     "react": "^18.3.1",
     "react-datepicker": "^8.4.0",
@@ -368,65 +348,6 @@
     "@modelcontextprotocol/server-filesystem": "^2025.7.29",
     "@modelcontextprotocol/server-memory": "^2025.8.4",
     "@modelcontextprotocol/server-sequential-thinking": "^2025.7.1"
-  },
-  "devDependencies": {
-    "@babel/eslint-parser": "^7.28.0",
-    "@babel/plugin-transform-class-properties": "^7.27.1",
-    "@babel/preset-env": "^7.28.3",
-    "@babel/preset-react": "^7.27.1",
-    "@babel/preset-typescript": "^7.27.1",
-    "@esbuild/linux-x64": "^0.25.9",
-    "@eslint/eslintrc": "^3.3.1",
-    "@eslint/js": "^9.34.0",
-    "@netlify/plugin-nextjs": "^5.12.1",
-    "@next/bundle-analyzer": "^15.5.2",
-    "@next/eslint-plugin-next": "^14.2.32",
-    "@rollup/rollup-linux-x64-gnu": "^4.50.0",
-    "@tailwindcss/aspect-ratio": "^0.4.2",
-    "@tailwindcss/forms": "^0.5.10",
-    "@tailwindcss/typography": "^0.5.16",
-    "@testing-library/jest-dom": "^6.6.4",
-    "@testing-library/react": "^16.3.0",
-    "@types/node": "^20.0.0",
-    "@types/node-cron": "^3.0.11",
-    "@types/react": "^18.0.0",
-    "@types/react-dom": "^18.0.0",
-    "@typescript-eslint/eslint-plugin": "^8.39.0",
-    "@typescript-eslint/parser": "^8.39.0",
-    "autoprefixer": "^10.4.16",
-    "eslint": "^8.57.0",
-    "eslint-config-next": "^14.0.0",
-    "@types/jest": "^29.5.8",
-    "@types/next": "^8.0.7",
-    "@types/node": "^18.0.0",
-    "@types/node-cron": "^3.0.11",
-    "@types/react": "^18.3.24",
-    "@types/react-dom": "^18.3.7",
-    "@types/react-router-dom": "^5.3.3",
-    "@typescript-eslint/eslint-plugin": "^8.42.0",
-    "@typescript-eslint/parser": "^8.42.0",
-    "@vitejs/plugin-react": "^5.0.2",
-    "autoprefixer": "^10.4.21",
-    "concurrently": "^8.2.2",
-    "cssnano": "^7.1.1",
-    "d3-color": "^3.1.0",
-    "d3-format": "^3.1.0",
-    "d3-path": "^3.1.0",
-    "d3-time-format": "^4.1.0",
-    "esbuild": "^0.25.9",
-    "eslint": "^9.34.0",
-    "eslint-config-next": "^14.2.32",
-    "eslint-formatter-junit": "^8.40.0",
-    "eslint-plugin-jsx-a11y": "^6.10.2",
-    "eslint-plugin-react": "^7.37.5",
-    "eslint-plugin-react-hooks": "^5.2.0",
-    "eslint-plugin-react-refresh": "^0.4.20",
-    "globals": "^16.3.0",
-    "husky": "^9.1.7",
-    "postcss": "^8.4.32",
-    "tailwindcss": "^3.4.0",
-    "ts-node": "^10.9.2",
-    "typescript": "^5.0.0"
   },
   "devDependencies": {
     "identity-obj-proxy": "^3.0.0",
